+++ airbyte-integrations/bases/source-acceptance-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.56
+Add test case in `TestDiscovery` and `TestConnection` to assert `additionalProperties` fields are set to true if they are declared [#14878](https://github.com/airbytehq/airbyte/pull/14878/).
+
+## 0.1.55
+Add test case in `TestDiscovery` to assert `supported_sync_modes` stream field in catalog is set and not empty.
+
 ## 0.1.54
 Fixed `AirbyteTraceMessage` test case to make connectors fail more reliably.